data_generation/iot_data_generation/requirements.txt
paho-mqtt
pydantic[dotenv]
sqlalchemy
asyncio
pydantic-settings
mysql-connector